+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Procedure when bumping the version number:
 ### Updated
 
 - Documentation for `time_zone` config option
+- When connecting to a room using `n` in the room list, the cursor now moves to that room
 
 ## v0.8.3 - 2024-05-20