+++ CHANGELOG.md
@@ -11,10 +11,11 @@ Procedure when bumping the version number:
 4. Run `cargo run help-config > CONFIG.md`
 5. Commit with message `Bump version to X.Y.Z`
 6. Create tag named `vX.Y.Z`
-7. Fast-forward branch `latest`
-8. Push `master`, `latest` and the new tag
+7. Push `master` and the new tag
 
 ## Unreleased
+
+## v0.9.0 - 2025-02-23
 
 ### Added